tests: add a x86_64 syscall resumption demo
While the kernel auditsc infrastructure is excellent for logging
system calls, it is useful to provide a soft-enforcement facility
for seccomp filter to act as a policy learning situation.

resumption.c provides a simple example that logs all the syscalls it sees,
including argument registers.  Since it is in-process, it could even log the
data of those arguments. This approach could be used to allow a slightly larger
set of system calls that are considered "safe" to be allowed via a thunk-only
until a final ruleset can be determined.

(Even if seccomp filter supported a SECCOMP_RET_LOG target, it'd need to grow
 a means to communicate its data to the userland process, etc.)

There is a lot of room for refinement, but it's nice to see that it isn't much
work to implement this for one arch (ignoring arch checking, arg checking, etc.