sigsegv: add a forced sigsegv example
This example shows a means to trigger a SIGSEGV after processing a system call
in SIGSYS while keeping all register intact.  It does so by mapping the
enclosing page PROT_NONE such that the sigreturn will fault.

This means it is unsafe to use code that may be adjacent to the original call
when performing the mprotect.  If not, the SIGSEGV will be triggered early.
Additionally, any SIGSEGV handlers may result in a double fault if they rely
on now inaccessible code.

This is useful for certain debugging and crash dumping frameworks which have
trouble unrolling stack frames, e.g., through a SIGSYS.  There are some other
potential uses, but that is the most obvious.