tests: invoke time vsyscall directly
The invocation of the time vsyscall produces a segfault when
built using a toolchain that produces position independent
executables by default. This is because the inline asm produces
an RIP-relative call whose offset does not get adjusted by the
dynamic linker.

Fix this by eliminating the vsyscall wrapper and defining
vsyscall_time to be a pointer to the fixed vsyscall address.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>